+++ Light Vox Engine/GraphicsCore.cpp
@@ -596,7 +596,7 @@ inline HRESULT GraphicsCore::InitInputShaderResources()
 		nullSrvDesc.Texture2D.MostDetailedMip = 0;
 		nullSrvDesc.Texture2D.ResourceMinLODClamp = 0.0f;
 
-        for (unsigned int i = 0; i < nullSrvCount; ++i)
+        for (size_t i = 0; i < nullSrvCount; ++i)
         {
 		    device->CreateShaderResourceView(nullptr, &nullSrvDesc, cbvSrvHandle);
 		    cbvSrvHandle.Offset(cbvSrvDescriptorSize);
